Prog_Contr_Mixup/requirements.txt
absl-py==0.13.0
astor==0.8.1
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
clang==5.0
click==8.0.1
configparser==5.0.2
cycler==0.10.0
dataclasses==0.8
diffdist==0.1
dirty-cat==0.2.0
docker-pycreds==0.4.0
faiss==1.5.3
faiss-gpu==1.7.1.post3
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.20
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.1.0
idna==2.10
importlib-metadata==4.6.0
install==1.3.4
joblib==1.0.1
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.1.5
pathtools==0.1.2
Pillow==8.2.0
progress==1.6
promise==2.3
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynvml==11.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-metric-learning==0.9.99
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
sentry-sdk==1.3.1
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
smmap==4.0.0
subprocess32==3.5.4
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0
tqdm==4.61.1
typing-extensions==3.7.4.3
urllib3==1.26.6
wandb==0.12.1
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
